requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.0
matplotlib==3.8.2
plotly==5.17.0
sympy==1.12
numpy==1.26.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
redis==5.0.1
